+++ app/(auth)/_components/easy-login.tsx
@@ -1,0 +1,3 @@
+export default function EasyLogin() {
+  return <h3>간편</h3>;
+}
